Move hashdict into new fontTools.misc.dictTools
